requirements.txt
Django>=5.2,<6.0
djangorestframework>=3.16,<4.0
python-dotenv>=1.0,<2.0
dj-database-url>=3.0,<4.0
celery>=5.6,<6.0
requests>=2.31,<3.0
yfinance>=0.2.40,<1.0
pandas>=2.0,<3.0